Suppression de certains fichiers inutiles en présence de Stéphane Galland
Début du remplacement des "float" par des "double" --> rencontre de
quelques méthodes sujet à questionnement

+++ core/math/src/main/java/org/arakhne/afc/math/MathConstants.java
@@ -82,7 +82,7 @@ public interface MathConstants {
 
 	/** Square root of 2
 	 */
-	public static final double SQRT_2 = (float)Math.sqrt(2.);
+	public static final double SQRT_2 = (double)Math.sqrt(2.);
 
 	/**
 	 * Max sweeps in the Jacoby's algorithms.
