+++ core/math/src/main/java/org/arakhne/afc/math/continous/object2d/Point2f.java
@@ -60,7 +60,7 @@ public class Point2f extends Tuple2f<Point2D> implements Point2D {
 	/**
 	 * @param tuple is the tuple to copy.
 	 */
-	public Point2f(float[] tuple) {
+	public Point2f(double[] tuple) {
 		super(tuple);
 	}
 
@@ -77,7 +77,7 @@ public class Point2f extends Tuple2f<Point2D> implements Point2D {
 	 * @param y
 	 */
 	public Point2f(float x, float y) {
-		super(x,y);
+		super((double)x,(double)y);
 	}
 
 	/**
@@ -85,7 +85,7 @@ public class Point2f extends Tuple2f<Point2D> implements Point2D {
 	 * @param y
 	 */
 	public Point2f(double x, double y) {
-		super((float)x,(float)y);
+		super(x,y);
 	}
 
 	/**
@@ -107,8 +107,8 @@ public class Point2f extends Tuple2f<Point2D> implements Point2D {
 	 * {@inheritDoc}
 	 */
 	@Override
-	public float distanceSquared(Point2D p1) {
-	      float dx, dy;
+	public double distanceSquared(Point2D p1) {
+		double dx, dy;
 	      dx = this.x-p1.getX();  
 	      dy = this.y-p1.getY();
 	      return (dx*dx+dy*dy);
@@ -118,18 +118,18 @@ public class Point2f extends Tuple2f<Point2D> implements Point2D {
 	 * {@inheritDoc}
 	 */
 	@Override
-	public float distance(Point2D p1) {
-	      float  dx, dy;
+	public double distance(Point2D p1) {
+		double  dx, dy;
 	      dx = this.x-p1.getX();  
 	      dy = this.y-p1.getY();
-	      return (float)Math.sqrt(dx*dx+dy*dy);
+	      return (double)Math.sqrt(dx*dx+dy*dy);
 	}
 
 	/**
 	 * {@inheritDoc}
 	 */
 	@Override
-	public float distanceL1(Point2D p1) {
+	public double distanceL1(Point2D p1) {
 	      return (Math.abs(this.x-p1.getX()) + Math.abs(this.y-p1.getY()));
 	}
 
@@ -137,7 +137,7 @@ public class Point2f extends Tuple2f<Point2D> implements Point2D {
 	 * {@inheritDoc}
 	 */
 	@Override
-	public float distanceLinf(Point2D p1) {
+	public double distanceLinf(Point2D p1) {
 	      return Math.max( Math.abs(this.x-p1.getX()), Math.abs(this.y-p1.getY()));
 	}
 
@@ -166,7 +166,7 @@ public class Point2f extends Tuple2f<Point2D> implements Point2D {
 	}
 
 	@Override
-	public void scaleAdd(float s, Vector2D t1, Point2D t2) {
+	public void scaleAdd(double s, Vector2D t1, Point2D t2) {
 		this.x = s * t1.getX() + t2.getX();
 		this.y = s * t1.getY() + t2.getY();
 	}
@@ -178,7 +178,7 @@ public class Point2f extends Tuple2f<Point2D> implements Point2D {
 	}
 
 	@Override
-	public void scaleAdd(float s, Point2D t1, Vector2D t2) {
+	public void scaleAdd(double s, Point2D t1, Vector2D t2) {
 		this.x = s * t1.getX() + t2.getX();
 		this.y = s * t1.getY() + t2.getY();
 	}
@@ -190,7 +190,7 @@ public class Point2f extends Tuple2f<Point2D> implements Point2D {
 	}
 
 	@Override
-	public void scaleAdd(float s, Vector2D t1) {
+	public void scaleAdd(double s, Vector2D t1) {
 		this.x = s * this.x + t1.getX();
 		this.y = s * this.y + t1.getY();
 	}
